pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.poetry]
authors = ["Frederik Aalund <fpa@sbtinstruments.com>"]
description = ""
name = "wright"
version = "0.1.0"

[tool.poetry.dependencies]
PyQt5 = {version = "^5.15.5", optional = true}
qasync = {version = "^0.22.0", optional = true}
QtAwesome = {version = "^1.1.1", optional = true}
"RPi.GPIO" = "^0.7.0"
anyio = "^3.1"
anyio-serial = "^0.1.5"
cryptography = "^3.4.6"
py3tftp = "^1.2.3"
pydantic = "^1.9.0"
python = "^3.9"
smbus2 = "^0.4.1"
typer = "^0.3.2"
libconf = "^2.0.1"
asyncssh = "^2.7.2"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.8.0"
mypy = "^0.812"
pydocstyle = "^6.0.0"
pylint = "^2.7.4"
rope = "^0.18.0"

[tool.poetry.extras]
"gui" = ["PyQt5", "qasync", "QtAwesome"]

[tool.poetry.scripts]
shipyard-gui = 'wright.gui:gui'

[tool.black]
target-version = ['py38']

[tool.isort]
profile = 'black'

[tool.pylint.master]
# i,j,k: You can use this for an integer index in a loop.
# fs: We use this in the tests to get a pyfakefs file system reference.
#     Since pytest's dependency injection is name based, we have to use
#     this exact argument name.
# tg: You can use this for an `anyio.TaskGroup`.
# io: You can use this for `typing.BinaryIO` (and friends).
# T,U,V: For `TypeVar`s
good-names = 'i,j,k,fs,tg,io,T,U,V' # Default is 'i,j,k,ex,Run,_'
# Increase the default values a bit. Yes, this may hurt maintainability
# ever so slightly but it increases the speed of development.
max-args = '7' # Default is 5
max-attributes = '12' # Default is 7
# Match black's default max line length
max-line-length = '88' # Default is 100
# Ignore the following non-idiomatic modules
ignored-modules = 'RPi'

[tool.pylint.similarities]
# Imports are often duplicated. If two different files use a lot of
# the same imports, there is just no way around it. Therefore, we
# simply ignore import statements completely when we look for duplicate
# code.
ignore-imports = 'y'

[tool.pylint.'messages control']
# Note that there is an ongoing discussion about, the current
# pylint defaults:
#     https://github.com/PyCQA/pylint/issues/3512
#
# It's also interesting to note that the pylint authors disable
# some checks for the official pylint repo:
#     https://github.com/PyCQA/pylint/blob/master/pylintrc#L56

### IF YOU ENABLE A PYLINT CHECK THEN EXPLAIN WHY BELOW
enable = '''
'''

### IF YOU DISABLE A PYLINT CHECK THEN EXPLAIN WHY BELOW
#
# fixme: We use "TODO: " to note areas that we can improve.
#     It's nice to have this directly in the code in a way that
#     we can easily search for.
#     Related to: https://github.com/PyCQA/pylint/issues/2874
#
# missing-module-docstring,
# missing-class-docstring,
# missing-function-docstring: pydocstyle handles all this
#     See: https://pylint.readthedocs.io/en/latest/faq.html#i-am-using-another-popular-linter-alongside-pylint-which-messages-should-i-disable-to-avoid-duplicates
disable = '''
  fixme,
  missing-module-docstring,
  missing-class-docstring,
  missing-function-docstring,
'''
